# Invalidate hub-wide caches on deletions and overwrites