@teh-cmc 2024-09-27
> that's pretty much what you mean when complaining about the lack of secondary caches, right?

A real secondary cache would be aware of query semantics, and any shadowing introduced by these semantics, which is way more powerful and capable than working with row-ids directly.

E.g.:
* it would know to evict temporal data that is shadowed by newly logged static data. 
* it would know to evict data that is using latest-at semantics if the same index was relogged to.
* it would know to evict data if its index is shadowed by a (possibly recursive) clear.
* etc.

None of those are possible when working with row-ids directly.