@Wumpf 2024-09-27
`query awareness` is the keyword I was missing for my mental model, makes sense, thanks. Given what visualizer queries look like right now it sound quite challenging.
I figure a good first/next step would be to have more unified mechanism to store all data sources of a given query-entry (i.e. a list of rowids potentially from several stores (rowids are unique so the several store bit doesn't technically matter)) 🤔. Unfortunately that's not really enough since fallbacks may be generated using heuristic output that may change every frame. In that sense we can't make these caches "pure" secondary caches.